Ledger.Tx.Constraints

Description

Constraints for transactions

Documentation

This module defines TxConstraints, a list of constraints on transactions. To construct a value of TxConstraints use the mustPayToTheScriptWithDatumHash, mustSpendAtLeast, etc functions. Once we have a TxConstraints value it can be used both to generate a transaction that satisfies the constraints (off-chain, using mkTx) and to check whether a given pending transaction meets the constraints (on-chain, using checkScriptContext, checkScriptContext).

data TxConstraints i o

Restrictions placed on the allocation of funds to outputs of transactions.

Constructors

TxConstraints

Fields

txConstraints :: [TxConstraint]

txConstraintFuns :: TxConstraintFuns

txOwnInputs :: [ScriptInputConstraint i]

txOwnOutputs :: [ScriptOutputConstraint o]

data TxConstraint

Constraints on transactions that want to spend script outputs

Constructors

MustIncludeDatumInTxWithHash DatumHash Datum
The provided DatumHash and Datum must be included in the transaction body. Like MustIncludeDatumInTx, but useful when you already have a DatumHash and want to make sure that is is the actual hash of the Datum.

MustIncludeDatumInTx Datum
Like MustHashDatum, but the hash of the Datum is computed automatically.

MustValidateInTimeRange !(ValidityInterval POSIXTime)
The transaction's validity range must be set with the given POSIXTimeRange.

MustBeSignedBy PaymentPubKeyHash
The transaction must add the given PaymentPubKeyHash in its signatories.

MustSpendAtLeast Value
The sum of the transaction's input Values must be at least as much as the given Value.

MustProduceAtLeast Value
The sum of the transaction's output Values must be at least as much as the given Value.

MustSpendPubKeyOutput TxOutRef
The transaction must spend the given unspent transaction public key output.

MustSpendScriptOutput TxOutRef Redeemer (Maybe TxOutRef)
The transaction must spend the given unspent transaction script output.

MustUseOutputAsCollateral TxOutRef
The transaction must include the utxo as collateral input.

MustReferenceOutput TxOutRef
The transaction must reference (not spend) the given unspent transaction output.

MustMintValue MintingPolicyHash Redeemer TokenName Integer (Maybe TxOutRef)
The transaction must mint the given token and amount.

MustPayToAddress Address (Maybe (TxOutDatum Datum)) (Maybe ScriptHash) Value
The transaction must create a transaction output.

MustSatisfyAnyOf [[TxConstraint]]
The transaction must satisfy constraints given as an alternative of conjuctions (DNF), that is `check (MustSatisfyAnyOf xs) = any (all check) xs`

data ScriptInputConstraint a

Constraint which specifies that the transaction must spend a transaction output from a target script.

Constructors

ScriptInputConstraint

Fields

icRedeemer :: a

The typed Redeemer to be used with the target script

icTxOutRef :: TxOutRef

The UTXO to be spent by the target script

icReferenceTxOutRef :: Maybe TxOutRef

Optionally use a reference script as witness

data ScriptOutputConstraint a

Constructors

ScriptOutputConstraint

Fields

ocDatum :: TxOutDatum a

Typed datum to be used with the target script

ocValue :: Value

ocReferenceScriptHash :: Maybe ScriptHash

Defining constraints

mustPayToTheScriptWithDatumHash :: o -> Value -> TxConstraints i o

mustPayToTheScriptWithDatumHash d v locks the value v with a script alongside a datum d which is included in the transaction body.

If used in OffChain, this constraint creates a script output with dt and vl and adds dt in the transaction's datum witness set. The script address is derived from the typed validator that is provided in the ScriptLookups with typedValidatorLookups.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the new script transaction output with dt and vt is part of the transaction's outputs.

mustPayToTheScriptWithDatumInTx :: o -> Value -> TxConstraints i o

mustPayToTheScriptWithInlineDatum :: o -> Value -> TxConstraints i o

mustPayToAddress :: Address -> Value -> TxConstraints i o

mustPayToAddress addr v locks the value v at the given address addr.

If used in OffChain, this constraint creates a script output with addr and v.

If used in OnChain, this constraint verifies that the script transaction output with addr and v is part of the transaction's outputs.

mustPayToAddressWithDatumHash :: Address -> Datum -> Value -> TxConstraints i o

mustPayToAddress addr d v locks the value v at the given address addr alonside a datum d.

If used in OffChain, this constraint creates a script output with addr, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with addr, d and v is part of the transaction's outputs.

mustPayToAddressWithDatumInTx :: Address -> Datum -> Value -> TxConstraints i o

mustPayToAddressWithDatumInTx addr d v locks the value v at the given address addr alonside a datum d.

If used in OffChain, this constraint creates a script output with addr, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with addr, d and v as part of the transaction's outputs.

mustPayToAddressWithInlineDatum :: Address -> Datum -> Value -> TxConstraints i o

mustPayToAddressWithInlineDatum vh d v is the same as mustPayToAddress, but with an inline datum.

mustPayToAddressWithReferenceScript :: Address -> ScriptHash -> Maybe (TxOutDatum Datum) -> Value -> TxConstraints i o

mustPayToAddressWithReferenceScript addr scriptHash d v creates a transaction output with an reference script. This allows the script to be used as a reference script.

If used in OffChain, this constraint creates an output with addr, scriptHash, d and v and maybe adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the transaction output with addr, scriptHash, d and v is part of the transaction's outputs.

mustPayToAddressWithReferenceValidator :: Address -> ValidatorHash -> Maybe (TxOutDatum Datum) -> Value -> TxConstraints i o

mustPayToAddressWithReferenceValidator is a helper that calls mustPayToAddressWithReferenceScript.

mustPayToAddressWithReferenceMintingPolicy :: Address -> MintingPolicyHash -> Maybe (TxOutDatum Datum) -> Value -> TxConstraints i o

mustPayToAddressWithReferenceMintingPolicy is a helper that calls mustPayToAddressWithReferenceScript.

mustMintCurrency :: MintingPolicyHash -> TokenName -> Integer -> TxConstraints i o

Same as mustMintCurrencyWithRedeemer, but sets the redeemer to the unit redeemer.

mustMintCurrencyWithRedeemer :: MintingPolicyHash -> Redeemer -> TokenName -> Integer -> TxConstraints i o

Same as mustMintCurrencyWithRedeemerAndReference, but sets the reference to Nothing.

mustMintValue :: Value -> TxConstraints i o

Same as mustMintValueWithRedeemer, but sets the redeemer to the unit redeemer.

mustMintValueWithRedeemer :: Redeemer -> Value -> TxConstraints i o

Same as mustMintValueWithRedeemerAndReference, but sets the reference to Nothing.

mustSpendAtLeast :: Value -> TxConstraints i o

mustSpendAtLeast v requires the sum of the transaction's inputs value to be at least v.

If used in OffChain, this constraint checks if at least the given value is spent in the transaction. When the transaction is created, a DeclaredInputMismatch error is raised if it is not the case.

If used in OnChain, this constraint verifies that the sum of the transaction's inputs value to be at least v.

mustSpendPubKeyOutput :: TxOutRef -> TxConstraints i o

mustSpendPubKeyOutput utxo must spend the given unspent transaction public key output.

If used in OffChain, this constraint adds utxo as an input to the transaction. Information about this utxo must be provided in the ScriptLookups with unspentOutputs.

If several calls to mustSpendPubKeyOutput are performed for the same TxOutRef, only one instance of the constraint is kept when the transaction is created.

If used in OnChain, this constraint verifies that the transaction spends this utxo.

mustSpendOutputFromTheScript :: TxOutRef -> i -> TxConstraints i o

mustSpendOutputFromTheScript txOutRef red spends the transaction output txOutRef with a script address using the redeemer red.

If used in OffChain, this constraint spends a script output txOutRef with redeemer red. The script address is derived from the typed validator that is provided in the ScriptLookups with typedValidatorLookups.

If used in OnChain, this constraint verifies that the spend script transaction output with red is part of the transaction's inputs.

mustSpendScriptOutput :: TxOutRef -> Redeemer -> TxConstraints i o

mustSpendScriptOutput utxo red must spend the given unspent transaction script output.

If used in OffChain, this constraint adds utxo and red as an input to the transaction. Information about this utxo must be provided in the ScriptLookups with unspentOutputs. The validator must be either provided by unspentOutputs or through otherScript . The datum must be either provided by unspentOutputs or through otherData.

If several calls to mustSpendScriptOutput are performed for the same TxOutRef, if the two constraints have different redeemers, an error will be thrown when the transaction is created. Otherwise, only one instance of the constraint is kept. If combined with mustSpendScriptOutputWithReference for the same TxOutRef, see mustSpendScriptOutputWithReference.

If used in OnChain, this constraint verifies that the transaction spends this utxo.

mustSpendScriptOutputWithReference :: TxOutRef -> Redeemer -> TxOutRef -> TxConstraints i o

mustSpendScriptOutputWithReference utxo red refTxOutref must spend the given unspent transaction script output, using a script reference as witness.

If used in OffChain, this constraint adds utxo and red as an input to the transaction, and refTxOutref as reference input. Information about utxo and refTxOutref must be provided in the ScriptLookups with unspentOutputs. The datum must be either provided by unspentOutputs or through otherData.

If several calls to mustSpendScriptOutputWithReference are performed for the same TxOutRef, if the two constraints have different redeemers, or if the two constraints use a different TxOutRef as a TxOutRef, an error will be thrown when the transaction is created. Otherwise, only one instance of the constraint is kept.

If combined with mustSpendScriptOutput for the same TxOutRef, an error is throw if they have a different redeemer. Otherwise, only one instance of the mustSpendScriptOutputWithReference constraint is kept, the mustSpendScriptOutput constraints are ignored.

If used in OnChain, this constraint verifies that the transaction spends this utxo.

mustSpendScriptOutputWithMatchingDatumAndValue :: ValidatorHash -> (Datum -> Bool) -> (Value -> Bool) -> Redeemer -> TxConstraints i o

mustSpendScriptOutputWithMatchingDatumAndValue validatorHash datumPredicate valuePredicate redeemer must spend an output locked by the given validator script hash, which includes a Datum that matches the given datum predicate and a Value that matches the given value predicate.

If used in OffChain, this constraint checks that there's exactly one output that matches the requirements, and then adds this as an input to the transaction with the given redeemer.

The outputs that will be considered need to be privided in the ScriptLookups with unspentOutputs.

If used in OnChain, this constraint verifies that there's at least one input that matches the requirements.

mustUseOutputAsCollateral :: TxOutRef -> TxConstraints i o

mustUseOutputAsCollateral utxo must use the given unspent transaction output reference as collateral input.

If used in OffChain, this constraint adds utxo as a collateral input to the transaction.

In OnChain this constraint has no effect, since no information about collateral inputs is passed to the scripts.

mustReferenceOutput :: TxOutRef -> TxConstraints i o

mustReferenceOutput utxo must reference (not spend!) the given unspent transaction output reference.

If used in OffChain, this constraint adds utxo as a reference input to the transaction.

If used in OnChain, this constraint verifies that the transaction references this utxo.

mustValidateInSlotRange :: ValidityInterval Slot -> TxConstraints i o

mustValidateInSlotRange r requires the transaction's validity slot range to be contained in Slot range r.

If used in OffChain, this constraint sets the transaction's validity slot range to r.

If used in OnChain, this constraint verifies that the slot range r is entirely contained in the transaction's validity time range.

mustValidateInTimeRange :: ValidityInterval POSIXTime -> TxConstraints i o

mustValidateInTimeRange r requires the transaction's validity time range to be contained in POSIXTime range r.

If used in OffChain, this constraint sets the transaction's validity time range to r.

If used in OnChain, this constraint verifies that the time range r is entirely contained in the transaction's validity time range.

mustBeSignedBy :: PaymentPubKeyHash -> TxConstraints i o

mustBeSignedBy pk requires the transaction to be signed by the public key pk.

If used in OffChain, this constraint adds pk in the transaction's public key witness set.

If used in OnChain, this constraint verifies that pk is part of the transaction's public key witness set.

mustProduceAtLeast :: Value -> TxConstraints i o

mustProduceAtLeast v requires the sum of the transaction's outputs value to be at least v.

If used in OffChain, this constraint checks if at least the given value is produced in the transaction. When the transaction is created, a DeclaredOutputMismatch error is raised if it is not the case.

If used in OnChain, this constraint verifies that the sum of the transaction's outputs value to be at least v.

mustIncludeDatumInTxWithHash :: DatumHash -> Datum -> TxConstraints i o

mustIncludeDatumInTxWithHash dh d requires the transaction body to include the datum hash dh and actual datum d.

If used in OffChain, this constraint adds dh and d in the transaction's body.

If used in OnChain, this constraint verifies that dh and d are part of the transaction's body.

mustIncludeDatumInTx :: Datum -> TxConstraints i o

mustIncludeDatumInTx d requires the transaction body to include the datum d.

If used in OffChain, this constraint adds d in the transaction's body alongside it's hash (which is computed automatically).

If used in OnChain, this constraint verifies that d is part of the transaction's body.

mustSatisfyAnyOf :: [TxConstraints i o] -> TxConstraints i o

Must-pay constraints for specific types of addresses

mustPayToPubKey :: PaymentPubKeyHash -> Value -> TxConstraints i o

mustPayToPubKey pkh v is the same as mustPayToPubKeyAddressWithDatumHash, but without any staking key hash and datum.

mustPayToPubKeyAddress :: PaymentPubKeyHash -> StakingCredential -> Value -> TxConstraints i o

mustPayToPubKeyAddress pkh skh v is the same as mustPayToPubKeyAddressWithDatumHash, but without any datum.

mustPayToPubKeyWithDatumHash :: PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyWithDatumHash pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but without the staking key hash.

mustPayToPubKeyAddressWithDatumHash :: PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyAddressWithDatumHash pkh skh d v locks a transaction output with a public key address.

If used in OffChain, this constraint creates a public key output with pkh, skh, d and v and maybe adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the public key transaction output with pkh, skh, d and v is part of the transaction's outputs.

mustPayToPubKeyWithDatumInTx :: PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyWithDatumInTx pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but with an inline datum and without the staking key hash.

mustPayToPubKeyAddressWithDatumInTx :: PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyAddressWithDatumInTx pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but the datum is also added in the transaction body.

mustPayToPubKeyWithInlineDatum :: PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyWithInlineDatum pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but with an inline datum and without the staking key hash.

mustPayToPubKeyAddressWithInlineDatum :: PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayWithInlineInlineDatumToPubKeyAddress pkh d v is the same as mustPayToPubKeyAddressWithInlineDatum, but the datum is inline in the Tx.

mustPayToOtherScriptWithDatumHash :: ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptWithDatumHash vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but without the staking key hash.

mustPayToOtherScriptWithDatumInTx :: ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptWithDatumInTx vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but without the staking key hash.

mustPayToOtherScriptWithInlineDatum :: ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptWithInlineDatum vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but with an inline datum and without the staking key hash.

mustPayToOtherScriptAddressWithDatumHash :: ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddressWithDatumHash vh svh d v locks the value v with the given script hash vh alonside a datum d.

If used in OffChain, this constraint creates a script output with vh, svh, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with vh, svh, d and v is part of the transaction's outputs. For v, this means that the transactions output must be at least the given value. The output can contain more, or different tokens, but the requested value v must be present.

mustPayToOtherScriptAddressWithDatumInTx :: ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddressWithDatumInTx vh svh d v locks the value v with the given script hash vh alonside a datum d.

If used in OffChain, this constraint creates a script output with vh, svh, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with vh, svh, d and v is part of the transaction's outputs. For v, this means that the transactions output must be at least the given value. The output can contain more, or different tokens, but the requested value v must be present.

mustPayToOtherScriptAddressWithInlineDatum :: ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddressInlineDatum vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but with an inline datum.

Defining off-chain only constraints

collectFromPlutusV1Script :: Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

A set of constraints for a transaction that collects PlutusV1 script outputs from the address of the given validator script, using the same redeemer script for all outputs.

collectFromPlutusV1ScriptFilter :: (TxOutRef -> DecoratedTxOut -> Bool) -> Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

collectFromTheScriptFilter :: (TxOutRef -> DecoratedTxOut -> Bool) -> Map TxOutRef DecoratedTxOut -> i -> TxConstraints i o

Given the pay to script address of the Validator, collect from it all the outputs that match a predicate, using the RedeemerValue.

collectFromTheScript :: Map TxOutRef DecoratedTxOut -> i -> TxConstraints i o

A version of collectFromScript that selects all outputs at the address

collectFromPlutusV2Script :: Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

A set of constraints for a transaction that collects PlutusV2 script outputs from the address of the given validator script, using the same redeemer script for all outputs.

collectFromPlutusV2ScriptFilter :: (TxOutRef -> DecoratedTxOut -> Bool) -> Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

Queries on constraints

modifiesUtxoSet :: TxConstraints i o -> Bool

Check whether every transaction that satisfies the constraints has to modify the UTXO set.

isSatisfiable :: TxConstraints i o -> Bool

Are the constraints satisfiable?

Off-chain transaction generation

data UnbalancedTx

An unbalanced transaction. It needs to be balanced and signed before it can be submitted to the ledger. See note [Submitting transactions from Plutus contracts] in Wallet.

Constructors

UnbalancedEmulatorTx

Fields

unBalancedEmulatorTx :: Tx

unBalancedTxRequiredSignatories :: Set PaymentPubKeyHash

These are all the payment public keys that should be used to request the signatories from the user's wallet. The signatories are what is required to sign the transaction before submitting it to the blockchain. Transaction validation will fail if the transaction is not signed by the required wallet.

unBalancedTxUtxoIndex :: Map TxOutRef TxOut

Utxo lookups that are used for adding inputs to the UnbalancedTx. Simply refers to slTxOutputs of ScriptLookups.

UnbalancedCardanoTx

Fields

unBalancedCardanoBuildTx :: CardanoBuildTx

unBalancedTxUtxoIndex :: Map TxOutRef TxOut

Utxo lookups that are used for adding inputs to the UnbalancedTx. Simply refers to slTxOutputs of ScriptLookups.

data MkTxError

Constructors

ToCardanoError ToCardanoError

LedgerMkTxError MkTxError

mkTx :: (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => Params -> ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Either MkTxError UnbalancedTx

Turn a TxConstraints value into an unbalanced transaction that satisfies the constraints. To use this in a contract, see submitTxConstraints and related functions.

adjustUnbalancedTx :: PParams -> UnbalancedTx -> Either ToCardanoError ([Ada], UnbalancedTx)

Each transaction output should contain a minimum amount of Ada (this is a restriction on the real Cardano network).

Combining multiple typed scripts into one transaction

data SomeLookupsAndConstraints where

Some typed TxConstraints and the ScriptLookups needed to turn them into an UnbalancedTx.

Constructors

SomeLookupsAndConstraints :: forall a. (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> SomeLookupsAndConstraints

mkSomeTx :: Params -> [SomeLookupsAndConstraints] -> Either MkTxError UnbalancedTx

Given a list of SomeLookupsAndConstraints describing the constraints for several scripts, build a single transaction that runs all the scripts.

Lookups

data ScriptLookups a

Constructors

ScriptLookups

Fields

slTxOutputs :: Map TxOutRef DecoratedTxOut

Unspent outputs that the script may want to spend

slOtherScripts :: Map ScriptHash (Versioned Script)

Scripts other than "our script"

slOtherData :: Map DatumHash Datum

Datums that we might need

slPaymentPubKeyHashes :: Set PaymentPubKeyHash

Public keys that we might need

slTypedValidator :: Maybe (TypedValidator a)

The script instance with the typed validator hash & actual compiled program

slOwnPaymentPubKeyHash :: Maybe PaymentPubKeyHash

The contract's payment public key hash, used for depositing tokens etc.

slOwnStakingCredential :: Maybe StakingCredential

The contract's staking credentials (optional)

typedValidatorLookups :: TypedValidator a -> ScriptLookups a

A script lookups value with a script instance. For convenience this also includes the minting policy script that forwards all checks to the instance's validator.

If called multiple times, only the first typed validator is kept:

typedValidatorLookups tv1 <> typedValidatorLookups tv2 <> ...
    == typedValidatorLookups tv1

unspentOutputs :: Map TxOutRef DecoratedTxOut -> ScriptLookups a

A script lookups value that uses the map of unspent outputs to resolve input constraints.

mintingPolicy :: Versioned MintingPolicy -> ScriptLookups a

A script lookups value with a versioned minting policy script.

plutusV1MintingPolicy :: MintingPolicy -> ScriptLookups a

A script lookups value with a PlutusV1 minting policy script.

plutusV2MintingPolicy :: MintingPolicy -> ScriptLookups a

A script lookups value with a PlutusV2 minting policy script.

otherScript :: Versioned Validator -> ScriptLookups a

A script lookups value with a versioned validator script.

plutusV1OtherScript :: Validator -> ScriptLookups a

A script lookups value with a PlutusV1 validator script.

plutusV2OtherScript :: Validator -> ScriptLookups a

A script lookups value with a PlutusV2 validator script.

otherData :: Datum -> ScriptLookups a

A script lookups value with a datum.

paymentPubKey :: PaymentPubKey -> ScriptLookups a

A script lookups value with a payment public key

paymentPubKeyHash :: PaymentPubKeyHash -> ScriptLookups a

A script lookups value with a payment public key

ownPaymentPubKeyHash :: PaymentPubKeyHash -> ScriptLookups a

A script lookups value with a payment public key hash.

If called multiple times, only the first payment public key hash is kept:

ownPaymentPubKeyHash pkh1 <> ownPaymentPubKeyHash pkh2 <> ...
    == ownPaymentPubKeyHash pkh1

ownStakingCredential :: StakingCredential -> ScriptLookups a

A script lookups value with staking credentials.

If called multiple times, only the first staking credential is kept:

ownStakingCredential skh1 <> ownStakingCredential skh2 <> ...
    == ownStakingCredential skh1

Deprecated

mustPayToTheScript :: o -> Value -> TxConstraints i o

mustPayToAddressWithDatum :: Address -> Datum -> Value -> TxConstraints i o

mustPayWithDatumToPubKey :: PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayWithDatumToPubKeyAddress :: PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayWithDatumInTxToPubKey :: PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayWithDatumInTxToPubKeyAddress :: PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayWithInlineDatumToPubKey :: PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayWithInlineDatumToPubKeyAddress :: PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScript :: ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddress :: ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustValidateIn :: POSIXTimeRange -> TxConstraints i o

mustValidateIn r requires the transaction's validity time range to be contained in POSIXTimeRange r.

If used in OffChain, this constraint sets the transaction's validity time range to r.

If used in OnChain, this constraint verifies that the time range r is entirely contained in the transaction's validity time range.