Ledger.Tx.Constraints.TxConstraints

Description

Constraints for transactions

Documentation

data TxOutDatum datum

How tx outs datum are embedded in a a Tx

We do not use TxOutDatum from cardano-node to provide easier to handel type (we don't type witnesses) and to have a distinction at the type leve between constraints that require a Datum and constraints (like MustPayToOtherScript) with an optional datum (like MustPayToPubKeyAddress).

Constructors

TxOutDatumHash datum
A datum specified in a transaction output using only it's hash, i.e. the datum is not inlined nor is it added in the transaction body.

TxOutDatumInTx datum
A datum specified in a transaction output using it's hash, while also adding the actual datum in the transaction body.

TxOutDatumInline datum
A datum inlined in a transaction output. It is *not* added in the transaction body.

getTxOutDatum :: TxOutDatum d -> d

isTxOutDatumHash :: TxOutDatum d -> Bool

isTxOutDatumInTx :: TxOutDatum d -> Bool

isTxOutDatumInline :: TxOutDatum d -> Bool

data TxConstraint

Constraints on transactions that want to spend script outputs

Constructors

MustIncludeDatumInTxWithHash DatumHash Datum
The provided DatumHash and Datum must be included in the transaction body. Like MustIncludeDatumInTx, but useful when you already have a DatumHash and want to make sure that is is the actual hash of the Datum.

MustIncludeDatumInTx Datum
Like MustHashDatum, but the hash of the Datum is computed automatically.

MustValidateInTimeRange !(ValidityInterval POSIXTime)
The transaction's validity range must be set with the given POSIXTimeRange.

MustBeSignedBy PaymentPubKeyHash
The transaction must add the given PaymentPubKeyHash in its signatories.

MustSpendAtLeast Value
The sum of the transaction's input Values must be at least as much as the given Value.

MustProduceAtLeast Value
The sum of the transaction's output Values must be at least as much as the given Value.

MustSpendPubKeyOutput TxOutRef
The transaction must spend the given unspent transaction public key output.

MustSpendScriptOutput TxOutRef Redeemer (Maybe TxOutRef)
The transaction must spend the given unspent transaction script output.

MustUseOutputAsCollateral TxOutRef
The transaction must include the utxo as collateral input.

MustReferenceOutput TxOutRef
The transaction must reference (not spend) the given unspent transaction output.

MustMintValue MintingPolicyHash Redeemer TokenName Integer (Maybe TxOutRef)
The transaction must mint the given token and amount.

MustPayToAddress Address (Maybe (TxOutDatum Datum)) (Maybe ScriptHash) Value
The transaction must create a transaction output.

MustSatisfyAnyOf [[TxConstraint]]
The transaction must satisfy constraints given as an alternative of conjuctions (DNF), that is `check (MustSatisfyAnyOf xs) = any (all check) xs`

data TxConstraintFun

Constraints on transactions that contain functions. These don't support conversion to and from JSON.

Constructors

MustSpendScriptOutputWithMatchingDatumAndValue ValidatorHash (Datum -> Bool) (Value -> Bool) Redeemer
The transaction must spend a script output from the given script address which matches the Datum and Value predicates.

newtype TxConstraintFuns

Constructors

TxConstraintFuns [TxConstraintFun]

data ScriptInputConstraint a

Constraint which specifies that the transaction must spend a transaction output from a target script.

Constructors

ScriptInputConstraint

Fields

icRedeemer :: a

The typed Redeemer to be used with the target script

icTxOutRef :: TxOutRef

The UTXO to be spent by the target script

icReferenceTxOutRef :: Maybe TxOutRef

Optionally use a reference script as witness

mustSpendOutputFromTheScript :: TxOutRef -> i -> TxConstraints i o

mustSpendOutputFromTheScript txOutRef red spends the transaction output txOutRef with a script address using the redeemer red.

If used in OffChain, this constraint spends a script output txOutRef with redeemer red. The script address is derived from the typed validator that is provided in the ScriptLookups with typedValidatorLookups.

If used in OnChain, this constraint verifies that the spend script transaction output with red is part of the transaction's inputs.

mustSpendOutputFromTheReferencedScript :: TxOutRef -> i -> TxOutRef -> TxConstraints i o

mustSpendOutputFromTheReferencedScript txOutRef red ref spends the transaction output txOutRef with a script address using the redeemer red, using the reference script ref as a validator.

If used in OffChain, this constraint spends a script output txOutRef with redeemer red. The script address is derived from the typed validator that is provided in the ScriptLookups with typedValidatorLookups.

If used in OnChain, this constraint verifies that the spend script transaction output with red is part of the transaction's inputs.

data ScriptOutputConstraint a

Constructors

ScriptOutputConstraint

Fields

ocDatum :: TxOutDatum a

Typed datum to be used with the target script

ocValue :: Value

ocReferenceScriptHash :: Maybe ScriptHash

data TxConstraints i o

Restrictions placed on the allocation of funds to outputs of transactions.

Constructors

TxConstraints

Fields

txConstraints :: [TxConstraint]

txConstraintFuns :: TxConstraintFuns

txOwnInputs :: [ScriptInputConstraint i]

txOwnOutputs :: [ScriptOutputConstraint o]

type UntypedConstraints = TxConstraints BuiltinData BuiltinData

singleton :: TxConstraint -> TxConstraints i o

mustValidateIn :: forall i o. POSIXTimeRange -> TxConstraints i o

Deprecated: Please use mustValidateInTimeRange or mustValidateInSlotRange instead

mustValidateIn r requires the transaction's validity time range to be contained in POSIXTimeRange r.

If used in OffChain, this constraint sets the transaction's validity time range to r.

If used in OnChain, this constraint verifies that the time range r is entirely contained in the transaction's validity time range.

mustValidateInTimeRange :: forall i o. ValidityInterval POSIXTime -> TxConstraints i o

mustValidateInTimeRange r requires the transaction's validity time range to be contained in POSIXTime range r.

If used in OffChain, this constraint sets the transaction's validity time range to r.

If used in OnChain, this constraint verifies that the time range r is entirely contained in the transaction's validity time range.

mustValidateInSlotRange :: forall i o. ValidityInterval Slot -> TxConstraints i o

mustValidateInSlotRange r requires the transaction's validity slot range to be contained in Slot range r.

If used in OffChain, this constraint sets the transaction's validity slot range to r.

If used in OnChain, this constraint verifies that the slot range r is entirely contained in the transaction's validity time range.

mustBeSignedBy :: forall i o. PaymentPubKeyHash -> TxConstraints i o

mustBeSignedBy pk requires the transaction to be signed by the public key pk.

If used in OffChain, this constraint adds pk in the transaction's public key witness set.

If used in OnChain, this constraint verifies that pk is part of the transaction's public key witness set.

mustIncludeDatumInTxWithHash :: DatumHash -> Datum -> TxConstraints i o

mustIncludeDatumInTxWithHash dh d requires the transaction body to include the datum hash dh and actual datum d.

If used in OffChain, this constraint adds dh and d in the transaction's body.

If used in OnChain, this constraint verifies that dh and d are part of the transaction's body.

mustIncludeDatumInTx :: forall i o. Datum -> TxConstraints i o

mustIncludeDatumInTx d requires the transaction body to include the datum d.

If used in OffChain, this constraint adds d in the transaction's body alongside it's hash (which is computed automatically).

If used in OnChain, this constraint verifies that d is part of the transaction's body.

mustPayToTheScript :: o -> Value -> TxConstraints i o

Deprecated: Use mustPayToTheScriptWithDatumHash instead

mustPayToTheScriptWithDatumHash :: o -> Value -> TxConstraints i o

mustPayToTheScriptWithDatumHash d v locks the value v with a script alongside a datum d which is included in the transaction body.

If used in OffChain, this constraint creates a script output with dt and vl and adds dt in the transaction's datum witness set. The script address is derived from the typed validator that is provided in the ScriptLookups with typedValidatorLookups.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the new script transaction output with dt and vt is part of the transaction's outputs.

mustPayToTheScriptWithDatumInTx :: o -> Value -> TxConstraints i o

mustPayToTheScriptWithInlineDatum :: o -> Value -> TxConstraints i o

mustPayToTheScriptWithReferenceScript :: ScriptHash -> TxOutDatum o -> Value -> TxConstraints i o

mustPayToPubKey :: forall i o. PaymentPubKeyHash -> Value -> TxConstraints i o

mustPayToPubKey pkh v is the same as mustPayToPubKeyAddressWithDatumHash, but without any staking key hash and datum.

mustPayToPubKeyAddress :: forall i o. PaymentPubKeyHash -> StakingCredential -> Value -> TxConstraints i o

mustPayToPubKeyAddress pkh skh v is the same as mustPayToPubKeyAddressWithDatumHash, but without any datum.

mustPayWithDatumToPubKey :: forall i o. PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToPubKeyWithDatumHash instead

mustPayToPubKeyWithDatumHash :: forall i o. PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyWithDatumHash pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but without the staking key hash.

mustPayWithDatumInTxToPubKey :: forall i o. PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToPubKeyWithDatumInTx instead

mustPayToPubKeyWithDatumInTx :: forall i o. PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyWithDatumInTx pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but with an inline datum and without the staking key hash.

mustPayWithInlineDatumToPubKey :: forall i o. PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToPubKeyWithInlineDatum instead

mustPayToPubKeyWithInlineDatum :: forall i o. PaymentPubKeyHash -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyWithInlineDatum pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but with an inline datum and without the staking key hash.

mustPayWithDatumToPubKeyAddress :: forall i o. PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToPubKeyAddressWithDatumHash instead

mustPayToPubKeyAddressWithDatumHash :: forall i o. PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyAddressWithDatumHash pkh skh d v locks a transaction output with a public key address.

If used in OffChain, this constraint creates a public key output with pkh, skh, d and v and maybe adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the public key transaction output with pkh, skh, d and v is part of the transaction's outputs.

mustPayWithDatumInTxToPubKeyAddress :: forall i o. PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToPubKeyAddressWithDatumInTx instead

mustPayToPubKeyAddressWithDatumInTx :: forall i o. PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToPubKeyAddressWithDatumInTx pkh d v is the same as mustPayToPubKeyAddressWithDatumHash, but the datum is also added in the transaction body.

mustPayWithInlineDatumToPubKeyAddress :: forall i o. PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToPubKeyAddressWithInlineDatum instead

mustPayToPubKeyAddressWithInlineDatum :: forall i o. PaymentPubKeyHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayWithInlineInlineDatumToPubKeyAddress pkh d v is the same as mustPayToPubKeyAddressWithInlineDatum, but the datum is inline in the Tx.

mustPayToAddressWithReferenceValidator :: forall i o. Address -> ValidatorHash -> Maybe (TxOutDatum Datum) -> Value -> TxConstraints i o

mustPayToAddressWithReferenceValidator is a helper that calls mustPayToAddressWithReferenceScript.

mustPayToAddressWithReferenceMintingPolicy :: forall i o. Address -> MintingPolicyHash -> Maybe (TxOutDatum Datum) -> Value -> TxConstraints i o

mustPayToAddressWithReferenceMintingPolicy is a helper that calls mustPayToAddressWithReferenceScript.

mustPayToAddressWithReferenceScript :: forall i o. Address -> ScriptHash -> Maybe (TxOutDatum Datum) -> Value -> TxConstraints i o

mustPayToAddressWithReferenceScript addr scriptHash d v creates a transaction output with an reference script. This allows the script to be used as a reference script.

If used in OffChain, this constraint creates an output with addr, scriptHash, d and v and maybe adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the transaction output with addr, scriptHash, d and v is part of the transaction's outputs.

mustPayToOtherScript :: forall i o. ValidatorHash -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToOtherScriptWithDatumHash instead

mustPayToOtherScriptWithDatumHash :: forall i o. ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptWithDatumHash vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but without the staking key hash.

mustPayToOtherScriptWithDatumInTx :: forall i o. ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptWithDatumInTx vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but without the staking key hash.

mustPayToOtherScriptWithInlineDatum :: forall i o. ValidatorHash -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptWithInlineDatum vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but with an inline datum and without the staking key hash.

mustPayToOtherScriptAddress :: forall i o. ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToOtherScriptAddressWithDatumHash instead

mustPayToOtherScriptAddressWithDatumHash :: forall i o. ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddressWithDatumHash vh svh d v locks the value v with the given script hash vh alonside a datum d.

If used in OffChain, this constraint creates a script output with vh, svh, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with vh, svh, d and v is part of the transaction's outputs. For v, this means that the transactions output must be at least the given value. The output can contain more, or different tokens, but the requested value v must be present.

mustPayToOtherScriptAddressWithDatumInTx :: forall i o. ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddressWithDatumInTx vh svh d v locks the value v with the given script hash vh alonside a datum d.

If used in OffChain, this constraint creates a script output with vh, svh, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with vh, svh, d and v is part of the transaction's outputs. For v, this means that the transactions output must be at least the given value. The output can contain more, or different tokens, but the requested value v must be present.

mustPayToOtherScriptAddressWithInlineDatum :: forall i o. ValidatorHash -> StakingCredential -> Datum -> Value -> TxConstraints i o

mustPayToOtherScriptAddressInlineDatum vh d v is the same as mustPayToOtherScriptAddressWithDatumHash, but with an inline datum.

mustPayToAddress :: forall i o. Address -> Value -> TxConstraints i o

mustPayToAddress addr v locks the value v at the given address addr.

If used in OffChain, this constraint creates a script output with addr and v.

If used in OnChain, this constraint verifies that the script transaction output with addr and v is part of the transaction's outputs.

mustPayToAddressWithDatum :: forall i o. Address -> Datum -> Value -> TxConstraints i o

Deprecated: Use mustPayToAddressWithDatumHash instead

mustPayToAddressWithDatumHash :: forall i o. Address -> Datum -> Value -> TxConstraints i o

mustPayToAddress addr d v locks the value v at the given address addr alonside a datum d.

If used in OffChain, this constraint creates a script output with addr, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with addr, d and v is part of the transaction's outputs.

mustPayToAddressWithDatumInTx :: forall i o. Address -> Datum -> Value -> TxConstraints i o

mustPayToAddressWithDatumInTx addr d v locks the value v at the given address addr alonside a datum d.

If used in OffChain, this constraint creates a script output with addr, d and v and adds d in the transaction's datum witness set.

If used in OnChain, this constraint verifies that d is part of the datum witness set and that the script transaction output with addr, d and v as part of the transaction's outputs.

mustPayToAddressWithInlineDatum :: forall i o. Address -> Datum -> Value -> TxConstraints i o

mustPayToAddressWithInlineDatum vh d v is the same as mustPayToAddress, but with an inline datum.

mustMintValue :: forall i o. Value -> TxConstraints i o

Same as mustMintValueWithRedeemer, but sets the redeemer to the unit redeemer.

mustMintValueWithReference :: forall i o. TxOutRef -> Value -> TxConstraints i o

Same as mustMintValueWithRedeemerAndReference, but sets the redeemer to the unit redeemer.

mustMintValueWithRedeemer :: forall i o. Redeemer -> Value -> TxConstraints i o

Same as mustMintValueWithRedeemerAndReference, but sets the reference to Nothing.

mustMintValueWithRedeemerAndReference :: forall i o. Redeemer -> Maybe TxOutRef -> Value -> TxConstraints i o

Same as mustMintCurrencyWithRedeemerAndReference, but uses the minting policy hash, token name and amount provided by Value.

Note that we can derive the MintingPolicyHash from the Value's currency symbol.

mustMintCurrency :: forall i o. MintingPolicyHash -> TokenName -> Integer -> TxConstraints i o

Same as mustMintCurrencyWithRedeemer, but sets the redeemer to the unit redeemer.

mustMintCurrencyWithReference :: forall i o. TxOutRef -> MintingPolicyHash -> TokenName -> Integer -> TxConstraints i o

Same as mustMintCurrencyWithRedeemerAndReference, but sets the redeemer to the unit redeemer.

mustMintCurrencyWithRedeemer :: forall i o. MintingPolicyHash -> Redeemer -> TokenName -> Integer -> TxConstraints i o

Same as mustMintCurrencyWithRedeemerAndReference, but sets the reference to Nothing.

mustMintCurrencyWithRedeemerAndReference :: forall i o. Maybe TxOutRef -> MintingPolicyHash -> Redeemer -> TokenName -> Integer -> TxConstraints i o

mustMintCurrencyWithRedeemerAndReference mref mph r tn a creates the given amount a of the currency specified with mph, r and tn. The minting policy script can be specified with a reference script mref.

If used in OffChain, this constraint mints a currency using mref, mph, r, tn and a, adds mph in the transaction's minting policy witness set and adds r in the transaction's redeemer witness set. The minting policy must be provided in the ScriptLookups with typedValidatorLookups or plutusV1MintingPolicy.

If used in OnChain, this constraint verifies that the minted currenty mref, mph, tn and a is part of the transaction's minting information.

mustSpendAtLeast :: forall i o. Value -> TxConstraints i o

mustSpendAtLeast v requires the sum of the transaction's inputs value to be at least v.

If used in OffChain, this constraint checks if at least the given value is spent in the transaction. When the transaction is created, a DeclaredInputMismatch error is raised if it is not the case.

If used in OnChain, this constraint verifies that the sum of the transaction's inputs value to be at least v.

mustProduceAtLeast :: forall i o. Value -> TxConstraints i o

mustProduceAtLeast v requires the sum of the transaction's outputs value to be at least v.

If used in OffChain, this constraint checks if at least the given value is produced in the transaction. When the transaction is created, a DeclaredOutputMismatch error is raised if it is not the case.

If used in OnChain, this constraint verifies that the sum of the transaction's outputs value to be at least v.

mustSpendPubKeyOutput :: forall i o. TxOutRef -> TxConstraints i o

mustSpendPubKeyOutput utxo must spend the given unspent transaction public key output.

If used in OffChain, this constraint adds utxo as an input to the transaction. Information about this utxo must be provided in the ScriptLookups with unspentOutputs.

If several calls to mustSpendPubKeyOutput are performed for the same TxOutRef, only one instance of the constraint is kept when the transaction is created.

If used in OnChain, this constraint verifies that the transaction spends this utxo.

mustSpendScriptOutput :: forall i o. TxOutRef -> Redeemer -> TxConstraints i o

mustSpendScriptOutput utxo red must spend the given unspent transaction script output.

If used in OffChain, this constraint adds utxo and red as an input to the transaction. Information about this utxo must be provided in the ScriptLookups with unspentOutputs. The validator must be either provided by unspentOutputs or through otherScript . The datum must be either provided by unspentOutputs or through otherData.

If several calls to mustSpendScriptOutput are performed for the same TxOutRef, if the two constraints have different redeemers, an error will be thrown when the transaction is created. Otherwise, only one instance of the constraint is kept. If combined with mustSpendScriptOutputWithReference for the same TxOutRef, see mustSpendScriptOutputWithReference.

If used in OnChain, this constraint verifies that the transaction spends this utxo.

mustSpendScriptOutputWithReference :: TxOutRef -> Redeemer -> TxOutRef -> TxConstraints i o

mustSpendScriptOutputWithReference utxo red refTxOutref must spend the given unspent transaction script output, using a script reference as witness.

If used in OffChain, this constraint adds utxo and red as an input to the transaction, and refTxOutref as reference input. Information about utxo and refTxOutref must be provided in the ScriptLookups with unspentOutputs. The datum must be either provided by unspentOutputs or through otherData.

If several calls to mustSpendScriptOutputWithReference are performed for the same TxOutRef, if the two constraints have different redeemers, or if the two constraints use a different TxOutRef as a TxOutRef, an error will be thrown when the transaction is created. Otherwise, only one instance of the constraint is kept.

If combined with mustSpendScriptOutput for the same TxOutRef, an error is throw if they have a different redeemer. Otherwise, only one instance of the mustSpendScriptOutputWithReference constraint is kept, the mustSpendScriptOutput constraints are ignored.

If used in OnChain, this constraint verifies that the transaction spends this utxo.

mustSpendScriptOutputWithMatchingDatumAndValue :: forall i o. ValidatorHash -> (Datum -> Bool) -> (Value -> Bool) -> Redeemer -> TxConstraints i o

mustSpendScriptOutputWithMatchingDatumAndValue validatorHash datumPredicate valuePredicate redeemer must spend an output locked by the given validator script hash, which includes a Datum that matches the given datum predicate and a Value that matches the given value predicate.

If used in OffChain, this constraint checks that there's exactly one output that matches the requirements, and then adds this as an input to the transaction with the given redeemer.

The outputs that will be considered need to be privided in the ScriptLookups with unspentOutputs.

If used in OnChain, this constraint verifies that there's at least one input that matches the requirements.

mustUseOutputAsCollateral :: forall i o. TxOutRef -> TxConstraints i o

mustUseOutputAsCollateral utxo must use the given unspent transaction output reference as collateral input.

If used in OffChain, this constraint adds utxo as a collateral input to the transaction.

In OnChain this constraint has no effect, since no information about collateral inputs is passed to the scripts.

mustReferenceOutput :: forall i o. TxOutRef -> TxConstraints i o

mustReferenceOutput utxo must reference (not spend!) the given unspent transaction output reference.

If used in OffChain, this constraint adds utxo as a reference input to the transaction.

If used in OnChain, this constraint verifies that the transaction references this utxo.

mustSatisfyAnyOf :: forall i o. [TxConstraints i o] -> TxConstraints i o

isSatisfiable :: forall i o. TxConstraints i o -> Bool

Are the constraints satisfiable?

pubKeyPayments :: forall i o. TxConstraints i o -> [(PaymentPubKeyHash, Value)]

mustSpendAtLeastTotal :: forall i o. TxConstraints i o -> Value

The minimum Value that satisfies all MustSpendAtLeast constraints

mustProduceAtLeastTotal :: forall i o. TxConstraints i o -> Value

The minimum Value that satisfies all MustProduceAtLeast constraints

requiredSignatories :: forall i o. TxConstraints i o -> [PaymentPubKeyHash]

requiredMonetaryPolicies :: forall i o. TxConstraints i o -> [MintingPolicyHash]

requiredDatums :: forall i o. TxConstraints i o -> [Datum]

modifiesUtxoSet :: forall i o. TxConstraints i o -> Bool

Check whether every transaction that satisfies the constraints has to modify the UTXO set.

spendUtxosFromPlutusV1Script :: Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

A set of constraints for a transaction that collects PlutusV1 script outputs from the address of the given validator script, using the same redeemer script for all outputs.

spendUtxosFromPlutusV1ScriptFilter :: (TxOutRef -> DecoratedTxOut -> Bool) -> Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

spendUtxosFromTheScriptFilter :: forall i o. (TxOutRef -> DecoratedTxOut -> Bool) -> Map TxOutRef DecoratedTxOut -> i -> TxConstraints i o

Given the pay to script address of the Validator, collect from it all the outputs that match a predicate, using the RedeemerValue.

spendUtxosFromTheScript :: forall i o. Map TxOutRef DecoratedTxOut -> i -> TxConstraints i o

A version of spendUtxosFromScript that selects all outputs at the address

spendUtxosFromTheReferencedScript :: forall i o. Map TxOutRef DecoratedTxOut -> i -> TxOutRef -> TxConstraints i o

A version of spendUtxosFromScript that selects all outputs at the address

utxo the set of utxos we search into to find the one we want to spendsOutput ref the reference to the utxo that contains the reference script

spendUtxosFromPlutusV2Script :: Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints

A set of constraints for a transaction that collects PlutusV2 script outputs from the address of the given validator script, using the same redeemer script for all outputs.

spendUtxosFromPlutusV2ScriptFilter :: (TxOutRef -> DecoratedTxOut -> Bool) -> Map Address (Map TxOutRef DecoratedTxOut) -> Validator -> Redeemer -> UntypedConstraints