Ledger.Tx.Constraints.OffChain

Lookups

data ScriptLookups a

Constructors

ScriptLookups

Fields

slTxOutputs :: Map TxOutRef DecoratedTxOut

Unspent outputs that the script may want to spend

slOtherScripts :: Map ScriptHash (Versioned Script)

Scripts other than "our script"

slOtherData :: Map DatumHash Datum

Datums that we might need

slPaymentPubKeyHashes :: Set PaymentPubKeyHash

Public keys that we might need

slTypedValidator :: Maybe (TypedValidator a)

The script instance with the typed validator hash & actual compiled program

slOwnPaymentPubKeyHash :: Maybe PaymentPubKeyHash

The contract's payment public key hash, used for depositing tokens etc.

slOwnStakingCredential :: Maybe StakingCredential

The contract's staking credentials (optional)

typedValidatorLookups :: TypedValidator a -> ScriptLookups a

A script lookups value with a script instance. For convenience this also includes the minting policy script that forwards all checks to the instance's validator.

If called multiple times, only the first typed validator is kept:

typedValidatorLookups tv1 <> typedValidatorLookups tv2 <> ...
    == typedValidatorLookups tv1

generalise :: ScriptLookups a -> ScriptLookups Any

unspentOutputs :: Map TxOutRef DecoratedTxOut -> ScriptLookups a

A script lookups value that uses the map of unspent outputs to resolve input constraints.

mintingPolicy :: Versioned MintingPolicy -> ScriptLookups a

A script lookups value with a versioned minting policy script.

plutusV1MintingPolicy :: MintingPolicy -> ScriptLookups a

A script lookups value with a PlutusV1 minting policy script.

plutusV2MintingPolicy :: MintingPolicy -> ScriptLookups a

A script lookups value with a PlutusV2 minting policy script.

otherScript :: Versioned Validator -> ScriptLookups a

A script lookups value with a versioned validator script.

plutusV1OtherScript :: Validator -> ScriptLookups a

A script lookups value with a PlutusV1 validator script.

plutusV2OtherScript :: Validator -> ScriptLookups a

A script lookups value with a PlutusV2 validator script.

otherData :: Datum -> ScriptLookups a

A script lookups value with a datum.

ownPaymentPubKeyHash :: PaymentPubKeyHash -> ScriptLookups a

A script lookups value with a payment public key hash.

If called multiple times, only the first payment public key hash is kept:

ownPaymentPubKeyHash pkh1 <> ownPaymentPubKeyHash pkh2 <> ...
    == ownPaymentPubKeyHash pkh1

ownStakingCredential :: StakingCredential -> ScriptLookups a

A script lookups value with staking credentials.

If called multiple times, only the first staking credential is kept:

ownStakingCredential skh1 <> ownStakingCredential skh2 <> ...
    == ownStakingCredential skh1

paymentPubKey :: PaymentPubKey -> ScriptLookups a

A script lookups value with a payment public key

paymentPubKeyHash :: PaymentPubKeyHash -> ScriptLookups a

A script lookups value with a payment public key

Constraints resolution

data SomeLookupsAndConstraints where

Some typed TxConstraints and the ScriptLookups needed to turn them into an UnbalancedTx.

Constructors

SomeLookupsAndConstraints :: forall a. (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> SomeLookupsAndConstraints

data UnbalancedTx

An unbalanced transaction. It needs to be balanced and signed before it can be submitted to the ledger. See note [Submitting transactions from Plutus contracts] in Wallet.

Constructors

UnbalancedEmulatorTx

Fields

unBalancedEmulatorTx :: Tx

unBalancedTxRequiredSignatories :: Set PaymentPubKeyHash

These are all the payment public keys that should be used to request the signatories from the user's wallet. The signatories are what is required to sign the transaction before submitting it to the blockchain. Transaction validation will fail if the transaction is not signed by the required wallet.

unBalancedTxUtxoIndex :: Map TxOutRef TxOut

Utxo lookups that are used for adding inputs to the UnbalancedTx. Simply refers to slTxOutputs of ScriptLookups.

UnbalancedCardanoTx

Fields

unBalancedCardanoBuildTx :: CardanoBuildTx

unBalancedTxUtxoIndex :: Map TxOutRef TxOut

Utxo lookups that are used for adding inputs to the UnbalancedTx. Simply refers to slTxOutputs of ScriptLookups.

unBalancedTxTx :: UnbalancedTx -> Either CardanoBuildTx Tx

tx :: Traversal' UnbalancedTx CardanoBuildTx

txValidityRange :: Lens' CardanoBuildTx (TxValidityLowerBound BabbageEra, TxValidityUpperBound BabbageEra)

txOuts :: Lens' CardanoBuildTx [TxOut CtxTx BabbageEra]

utxoIndex :: Lens' UnbalancedTx (Map TxOutRef TxOut)

emptyUnbalancedTx :: Params -> UnbalancedTx

adjustUnbalancedTx :: PParams -> UnbalancedTx -> Either ToCardanoError ([Ada], UnbalancedTx)

Each transaction output should contain a minimum amount of Ada (this is a restriction on the real Cardano network).

data MkTxError

Constructors

ToCardanoError ToCardanoError

LedgerMkTxError MkTxError

mkTx :: (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => Params -> ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Either MkTxError UnbalancedTx

Turn a TxConstraints value into an unbalanced transaction that satisfies the constraints. To use this in a contract, see submitTxConstraints and related functions.

mkSomeTx :: Params -> [SomeLookupsAndConstraints] -> Either MkTxError UnbalancedTx

Given a list of SomeLookupsAndConstraints describing the constraints for several scripts, build a single transaction that runs all the scripts.