Ledger.Tx.Constraints.OffChain

Lookups

data ScriptLookups a

Constructors

ScriptLookups

Fields

slTxOutputs :: Map TxOutRef DecoratedTxOut

Unspent outputs that the script may want to spend

slOtherScripts :: Map ScriptHash (Versioned Script)

Scripts other than "our script"

slOtherData :: Map DatumHash Datum

Datums that we might need

slPaymentPubKeyHashes :: Set PaymentPubKeyHash

Public keys that we might need

slTypedValidator :: Maybe (TypedValidator a)

The script instance with the typed validator hash & actual compiled program

slOwnPaymentPubKeyHash :: Maybe PaymentPubKeyHash

The contract's payment public key hash, used for depositing tokens etc.

slOwnStakingCredential :: Maybe StakingCredential

The contract's staking credentials (optional)

typedValidatorLookups :: TypedValidator a -> ScriptLookups a

A script lookups value with a script instance. For convenience this also includes the minting policy script that forwards all checks to the instance's validator.

If called multiple times, only the first typed validator is kept:

typedValidatorLookups tv1 <> typedValidatorLookups tv2 <> ...
    == typedValidatorLookups tv1

generalise :: ScriptLookups a -> ScriptLookups Any

unspentOutputs :: Map TxOutRef DecoratedTxOut -> ScriptLookups a

A script lookups value that uses the map of unspent outputs to resolve input constraints.

mintingPolicy :: Versioned MintingPolicy -> ScriptLookups a

A script lookups value with a versioned minting policy script.

plutusV1MintingPolicy :: MintingPolicy -> ScriptLookups a

A script lookups value with a PlutusV1 minting policy script.

plutusV2MintingPolicy :: MintingPolicy -> ScriptLookups a

A script lookups value with a PlutusV2 minting policy script.

otherScript :: Versioned Validator -> ScriptLookups a

A script lookups value with a versioned validator script.

plutusV1OtherScript :: Validator -> ScriptLookups a

A script lookups value with a PlutusV1 validator script.

plutusV2OtherScript :: Validator -> ScriptLookups a

A script lookups value with a PlutusV2 validator script.

otherData :: Datum -> ScriptLookups a

A script lookups value with a datum.

paymentPubKey :: PaymentPubKey -> ScriptLookups a

A script lookups value with a payment public key

paymentPubKeyHash :: PaymentPubKeyHash -> ScriptLookups a

A script lookups value with a payment public key

Constraints resolution

data SomeLookupsAndConstraints where

Some typed TxConstraints and the ScriptLookups needed to turn them into an UnbalancedTx.

Constructors

SomeLookupsAndConstraints :: forall a. (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> SomeLookupsAndConstraints

data UnbalancedTx

An unbalanced transaction. It needs to be balanced and signed before it can be submitted to the ledger. See note [Submitting transactions from Plutus contracts] in Wallet.

Constructors

UnbalancedCardanoTx

Fields

unBalancedCardanoBuildTx :: CardanoBuildTx

unBalancedTxUtxoIndex :: UtxoIndex

Utxo lookups that are used for adding inputs to the UnbalancedTx. Simply refers to slTxOutputs of ScriptLookups.

tx :: Traversal' UnbalancedTx CardanoBuildTx

txInsCollateral :: Lens' CardanoBuildTx [TxIn]

txValidityRange :: Lens' CardanoBuildTx (TxValidityLowerBound BabbageEra, TxValidityUpperBound BabbageEra)

txOuts :: Lens' CardanoBuildTx [TxOut]

utxoIndex :: Lens' UnbalancedTx UtxoIndex

emptyUnbalancedTx :: Params -> UnbalancedTx

adjustUnbalancedTx :: PParams -> UnbalancedTx -> ([Lovelace], UnbalancedTx)

Each transaction output should contain a minimum amount of Ada (this is a restriction on the real Cardano network).

mkTx :: (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => Params -> ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Either MkTxError UnbalancedTx

Turn a TxConstraints value into an unbalanced transaction that satisfies the constraints. To use this in a contract, see submitTxConstraints and related functions.

mkTxWithParams :: (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => Params -> ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Either MkTxError UnbalancedTx

Turn a TxConstraints value into an unbalanced transaction that satisfies the constraints. To use this in a contract, see submitTxConstraints and related functions.

mkSomeTx :: Params -> [SomeLookupsAndConstraints] -> Either MkTxError UnbalancedTx

Given a list of SomeLookupsAndConstraints describing the constraints for several scripts, build a single transaction that runs all the scripts.

data MkTxError

Constructors

TypeCheckFailed ConnectionError

ToCardanoError ToCardanoError

TxOutRefNotFound TxOutRef

TxOutRefWrongType TxOutRef

TxOutRefNoReferenceScript TxOutRef

DatumNotFound DatumHash

DeclaredInputMismatch Value

DeclaredOutputMismatch Value

MintingPolicyNotFound MintingPolicyHash

ScriptHashNotFound ScriptHash

TypedValidatorMissing

DatumWrongHash DatumHash Datum

CannotSatisfyAny

NoMatchingOutputFound ValidatorHash

MultipleMatchingOutputsFound ValidatorHash

AmbiguousRedeemer TxOutRef [Redeemer]

AmbiguousReferenceScript TxOutRef [TxOutRef]

_TypeCheckFailed :: AsMkTxError r => Prism' r ConnectionError

_ToCardanoError :: AsMkTxError r => Prism' r ToCardanoError

_TxOutRefNotFound :: AsMkTxError r => Prism' r TxOutRef

_TxOutRefWrongType :: AsMkTxError r => Prism' r TxOutRef

_TxOutRefNoReferenceScript :: AsMkTxError r => Prism' r TxOutRef

_DatumNotFound :: AsMkTxError r => Prism' r DatumHash

_DeclaredInputMismatch :: AsMkTxError r => Prism' r Value

_MintingPolicyNotFound :: AsMkTxError r => Prism' r MintingPolicyHash

_ScriptHashNotFound :: AsMkTxError r => Prism' r ScriptHash

_TypedValidatorMissing :: AsMkTxError r => Prism' r ()

_DatumWrongHash :: AsMkTxError r => Prism' r (DatumHash, Datum)

_CannotSatisfyAny :: AsMkTxError r => Prism' r ()

_NoMatchingOutputFound :: AsMkTxError r => Prism' r ValidatorHash

_MultipleMatchingOutputsFound :: AsMkTxError r => Prism' r ValidatorHash

Internals exposed for testing

data ValueSpentBalances

The balances we track for computing the missing Value (if any) that needs to be added to the transaction. See note [Balance of value spent].

Constructors

ValueSpentBalances

Fields

vbsRequired :: Value

Required value spent by the transaction.

vbsProvided :: Value

Value provided by an input or output of the transaction.

provided :: Value -> ValueSpentBalances

required :: Value -> ValueSpentBalances

missingValueSpent :: ValueSpentBalances -> Value

data ConstraintProcessingState

Constructors

ConstraintProcessingState

Fields

cpsUnbalancedTx :: UnbalancedTx

The unbalanced transaction that we're building

cpsValueSpentBalancesInputs :: ValueSpentBalances

Balance of the values given and required for the transaction's inputs

cpsValueSpentBalancesOutputs :: ValueSpentBalances

Balance of the values produced and required for the transaction's outputs

cpsParams :: Params

unbalancedTx :: Lens' ConstraintProcessingState UnbalancedTx

valueSpentInputs :: Lens' ConstraintProcessingState ValueSpentBalances

valueSpentOutputs :: Lens' ConstraintProcessingState ValueSpentBalances

paramsL :: Lens' ConstraintProcessingState Params

processConstraintFun :: TxConstraintFun -> ReaderT (ScriptLookups a) (StateT ConstraintProcessingState (Except MkTxError)) ()

addOwnInput :: (MonadReader (ScriptLookups a) m, MonadError MkTxError m, FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => ScriptInputConstraint (RedeemerType a) -> m TxConstraint

Add a typed input, checking the type of the output it spends. Return the value of the spent output.

addOwnOutput :: (MonadReader (ScriptLookups a) m, MonadError MkTxError m, ToData (DatumType a)) => ScriptOutputConstraint (DatumType a) -> m TxConstraint

Convert a ScriptOutputConstraint into a TxConstraint.

updateUtxoIndex :: (MonadReader (ScriptLookups a) m, MonadState ConstraintProcessingState m, MonadError MkTxError m) => m ()

lookupTxOutRef :: TxOutRef -> ReaderT (ScriptLookups a) (StateT ConstraintProcessingState (Except MkTxError)) DecoratedTxOut

lookupMintingPolicy :: MintingPolicyHash -> ReaderT (ScriptLookups a) (StateT ConstraintProcessingState (Except MkTxError)) (Versioned MintingPolicy)

lookupScript :: (MonadReader (ScriptLookups a) m, MonadError MkTxError m) => ScriptHash -> m (Versioned Script)

lookupScriptAsReferenceScript :: Maybe ScriptHash -> ReaderT (ScriptLookups a) (StateT ConstraintProcessingState (Except MkTxError)) (ReferenceScript BabbageEra)

prepareConstraints :: (FromData (DatumType a), ToData (DatumType a), ToData (RedeemerType a)) => [ScriptInputConstraint (RedeemerType a)] -> [ScriptOutputConstraint (DatumType a)] -> [TxConstraint] -> ReaderT (ScriptLookups a) (StateT ConstraintProcessingState (Except MkTxError)) SortedConstraints

resolveScriptTxOut :: (MonadReader (ScriptLookups a) m, MonadError MkTxError m) => DecoratedTxOut -> m (Maybe (Versioned Validator, DatumWithOrigin, Value))

resolveScriptTxOutValidator :: (MonadReader (ScriptLookups a) m, MonadError MkTxError m) => DecoratedTxOut -> m (Maybe (Versioned Validator))

resolveScriptTxOutDatumAndValue :: (MonadReader (ScriptLookups a) m, MonadError MkTxError m) => DecoratedTxOut -> m (Maybe (DatumWithOrigin, Value))

data DatumWithOrigin

Constructors

DatumInTx

Fields

getDatum :: Datum

DatumInline

Fields

getDatum :: Datum

datumWitness :: DatumWithOrigin -> Maybe Datum

checkValueSpent :: (MonadReader (ScriptLookups a) m, MonadState ConstraintProcessingState m, MonadError MkTxError m) => m ()

data SortedConstraints

Constructors

MkSortedConstraints

Fields

rangeConstraints :: [POSIXTimeRange]

otherConstraints :: [TxConstraint]

initialState :: Params -> ConstraintProcessingState